refactor: 기존에 사용하지않는 파일 삭제
