+++ src/main/java/bowling/Main.java
@@ -1,6 +1,6 @@
 package bowling;
 
-import bowling.domain.game.Game2;
+import bowling.domain.game.Game;
 import bowling.domain.player.Player;
 import bowling.domain.view.InputView;
 import bowling.domain.view.ResultView;
@@ -9,7 +9,7 @@ public class Main {
     public static void main(String[] args) {
         Player player = new Player(InputView.inputPlayerName());
 
-        Game2 game = Game2.start();
+        Game game = Game.start();
         ResultView resultView = new ResultView();
 
         while(!game.isGameOver()) {
