+++ src/main/java/bowling/domain/frame/FrameResult.java
@@ -1,16 +1,16 @@
 package bowling.domain.frame;
 
-import bowling.domain.score.Score3;
-import bowling.domain.score.ScoreType2;
+import bowling.domain.score.Score;
+import bowling.domain.score.ScoreType;
 
 import java.util.List;
 
 public class FrameResult {
     private List<Integer> downPins;
-    private ScoreType2 scoreType;
-    private Score3 score;
+    private ScoreType scoreType;
+    private Score score;
 
-    public FrameResult(List<Integer> downPins, ScoreType2 scoreType, Score3 score) {
+    public FrameResult(List<Integer> downPins, ScoreType scoreType, Score score) {
         this.downPins = downPins;
         this.scoreType = scoreType;
         this.score = score;
@@ -20,11 +20,11 @@ public class FrameResult {
         return downPins;
     }
 
-    public ScoreType2 getScoreType() {
+    public ScoreType getScoreType() {
         return scoreType;
     }
 
-    public Score3 getScore() {
+    public Score getScore() {
         return score;
     }
 }
