--- src/main/java/bowling/domain/score/Miss.java
@@ -1,41 +1,0 @@
-package bowling.domain.score;
-
-import bowling.domain.point.Point;
-
-public class Miss implements Score2 {
-    private static final int BONUS_POINT = 10;
-    public static final String ERROR_INPUT = "잘못된 입력입니다.";
-    private Point point;
-
-    public Miss(Point point) {
-        if (point.getPoint() == BONUS_POINT) {
-            throw new IllegalArgumentException(ERROR_INPUT);
-        }
-        this.point = point;
-    }
-
-    @Override
-    public Score2 nextScore(Point point) {
-        return new Normal(point);
-    }
-
-    @Override
-    public String getScore() {
-        return point.toString();
-    }
-
-    @Override
-    public boolean isStrike() {
-        return false;
-    }
-
-    @Override
-    public boolean isSpare() {
-        return false;
-    }
-
-    @Override
-    public int getPoint() {
-        return point.getPoint();
-    }
-}
