+++ src/main/java/bowling/domain/score/ScoreType.java
@@ -1,5 +1,19 @@
 package bowling.domain.score;
 
-public enum  ScoreType {
-    STRIKE, SPARE, NORMAL, GUTTER, MISS, NONE
+public enum ScoreType {
+    STRIKE(2),
+    SPARE(1),
+    NORMAL(0),
+    MISS(0),
+    READY(0);
+
+    private int bonusRollCount;
+
+    ScoreType(int bonusRollCount) {
+        this.bonusRollCount = bonusRollCount;
+    }
+
+    public int getBonusRollCount() {
+        return bonusRollCount;
+    }
 }
