+++ src/test/java/bowling/domain/pin/NormalFramePinsTest.java
@@ -1,7 +1,6 @@
 package bowling.domain.pin;
 
-import bowling.domain.frame.NormalFrame;
-import bowling.domain.score.ScoreType2;
+import bowling.domain.score.ScoreType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -14,17 +13,17 @@ class NormalFramePinsTest {
     @DisplayName("미스 상태인지 테스트")
     @Test
     void miss() {
-        Pins2 pins = new NormalFramePins();
+        Pins pins = new NormalFramePins();
         pins.down(3);
         pins.down(4);
 
-        assertThat(pins.getScoreType()).isEqualTo(ScoreType2.MISS);
+        assertThat(pins.getScoreType()).isEqualTo(ScoreType.MISS);
     }
 
     @DisplayName("두번 이상 핀 테스트")
     @Test
     void normalFramePin_third() {
-        Pins2 pins = new NormalFramePins();
+        Pins pins = new NormalFramePins();
         pins.down(3);
         pins.down(4);
 
@@ -36,7 +35,7 @@ class NormalFramePinsTest {
     @DisplayName("스트라이크 이후 핀 테스트")
     @Test
     void normalFrame_strike_third() {
-        Pins2 pins = new NormalFramePins();
+        Pins pins = new NormalFramePins();
         pins.down(10);
 
         assertThatThrownBy(() -> {
@@ -47,7 +46,7 @@ class NormalFramePinsTest {
     @DisplayName("핀 10개 이상 발생시 에러")
     @Test
     void normalFrame_pin() {
-        Pins2 pins = new NormalFramePins();
+        Pins pins = new NormalFramePins();
 
         assertThatThrownBy(() -> {
             pins.down(11);
@@ -57,7 +56,7 @@ class NormalFramePinsTest {
     @DisplayName("핀 합계 유효성 테스트")
     @Test
     void normalFrame_sum_pin() {
-        Pins2 pins = new NormalFramePins();
+        Pins pins = new NormalFramePins();
         pins.down(5);
         assertThatThrownBy(() -> {
             pins.down(6);
