--- src/test/java/bowling/domain/score/StrikeTest.java
@@ -1,17 +1,0 @@
-package bowling.domain.score;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class StrikeTest {
-
-    @DisplayName("점수 넘김 테스트")
-    @Test
-    void nextScore() {
-        Strike strike = new Strike();
-        assertThat(strike.isStrike()).isTrue();
-        assertThat(strike.getScore()).isEqualTo("X");
-    }
-}
